Use new method to construct block context
The method does not accept a vmrunner but now builds the vmrunner itself
from a state and header allowing it to copy the state before building
the runner. This encapsulates the state copy releiving callers from
having to know that the state needs to be copied here.

Also removed VmRunnerAtHeader methods which were simply duplicating
NewEVMRunner functionality but with a different name.