Ensure copy of state used to generate block context (#1900)
* Use new method to construct block context

The method to create a block context has been changed and now
does not accept a vmrunner but builds the vmrunner itself
from a state and header allowing it to copy the state before building
the runner. This encapsulates the state copy relieving callers from
having to know that the state needs to be copied here.

This solves a problem we were having with incorrect gas estimation where
transactions were sharing an access list (see go-ethereum PR #21509) with
previous EVM executions, resulting in gas estimations that were too low.

Also removed VmRunnerAtHeader methods which were simply duplicating
NewEVMRunner functionality but with a different name.