Update core system contracts for new e2e tests (#1893)
The tests under ./e2e_test were running against the system
contracts at version celo-core-contracts-v3.rc0, this PR updates
that version to be core-contracts.v6.
(because that is the most recently deployed version)

Also the make rule 'prepare-system-contracts' was previously broken in that it
would not rebuild the system contracts when the version identifier was
changed. The version identifier is now held in a file and when it is
modified the contracts are rebuilt.

Also clarify the use of the MONOREPO_PATH variable, making clear that it is
not intended to be used to point to existing manually checked out monorepos.