Update monorepo-checkout commit
Updated to use the commit held in the 'monorepo_commit' file at the root
of this repo.

This means that the e2e tests from the monorepo will be run against the
same contract version as the e2e tests contained in this repo.

Also tidy up the checkout code, simplify checkout with a one liner and
try to reduce amount of data persisted to workspace.