Update CI config to improve build speed
This update reduces overall build time from about 22m to 15m, a 7m
saving or about a 32% imrovement.

The time taken for the unit-tests job to complete has been dropped from
around 19m to about 4m, this should be beneficial to developers by
providing a much shorter feedback loop.

These changes were made possible by taking steps to minimise the amount
of data downloaded and shared between jobs, and most pertinently
removing usages of workspaces to persist and retrieve intermediate build
artefacts. And in place of that relying more on caches and modifying
jobs so that they are able to rebuild those caches. Greater use of
caches was enabled by better understanding of how to reuse portions of
config via yaml anchors.