Move install go out of checkout monorepo
This should hopefully improve performance by taking more out of the
workspace. Workspace persisting and loading is slow and the price is
paid multiple times.

EG if we install go in checkout monorepo, then all jobs that depend on
checkout monorepo wait for go to be installed (or the cache to be
restored) and then they wait for that data to be persisted to the
workspace and then they need to wait to attach to the
workspace.

By doing go install in each job directly we simply wait for the go
install, which is hopfully less than the other threee steps.