Fix CI, update core contracts version
trianglesphere previously updated the executor image for the system
contracts to a node v12 version -
1557fe681d7dea8859ec73510b71cff15496b953.

I think this broke the build but becase the executor image was not part
of the cache key, the existing cache continued to be used, until very
recently, and now we see that the build fails because it needs node
v10.

This commit leaves the executor at node v12 but updates the core system
contracts version to be in line with master, and that versoin actually
works with node v12.