Ensure RoundStateDB is closed when core is stopped (#1996)
Prior to this change the RoundStateDB was never closed. This could
result in leaked goroutines from its internal leveldb instance.

This change caused the TestStartStopValidators e2e test to fail because
changing the rsdb open to be in the startup code, slowed down the start
method for core this caused the gossiping of the manually added peers
(done in the test after starting each previously stopped node )to happen
after the validator peer handler had removed all the peers resulting in
gossiping certs to no one. Thus leaving the network in a disconnected
state.

The fix was really hacky, but just to wait 250ms after starting a node
and hoping that the validator peer handler will have removed nodes by
that point and then add the peers and then wait a little for them to be
registered and then finally gossip to them.

Co-authored-by: Paul Lange <palango@users.noreply.github.com>