graphql: add EIP-1559 and celo gas handling
The GraphQL-Endpoint did not properly handle post-EIP1559 gas fees. To
get the correct results based on the respective baseFee, we have to take
a look at the used feeCurrency.

To get at least some test coverage for this, I added one EIP-1559
transaction to the test data. This is still far from a thorough coverage
of all relevant cases (different feeCurrency, the base fee is 0 in the
test, so bugs for proper values might be hidden).