Activate hard forks in tests (#2089)
* Prune test chains

* Add minimumGasPrice to mock engine

* Test fixes for TestChainConfig

* Switch SimulatedBackend to TestChainConfig

Without this, the MockSysContractCallCtx is not used everywhere because
Faker is not set on the chain config. Due to this, the gas prices are
not the same in all execution paths, leading to mismatches in the state
merkle root.

* Set TestChainConfig.ChainId to 1337

Which is the chain ID geth uses in the simulated backend.

* Set mocked gas price to Zero

Some tests assume a gas price of zero when submitting transactions. To
avoid additional changes, assume a default gas price of zero in when
using MockSysContractCallCtx.

* Update TestNewSimulatedBackend to use TestChainConfig

* graphql test: update hashes after gas price change

Now that the mocked gas price is zero, the tx hash changed.

* Fix SysCtx creating on genesis block

We can't use the parent in that case, so let's use the block itself.

* odr_test: Use MockSysContractCallCtx

We need to add a SysCtx, now that we enable Espresso.

* Make fake base fee configurable

Most tests expect a base fee of zero, but sometimes we need to test
against non-zero base fees. To make both work, allow configuring the
FakeBaseFee in the ChainConfig.

This also nicely separates the two different uses of the `Faker` flag:
* faked consensus (still via the `Faker` flag)
* faked base fees (formerly activated via `Faker`, now via `FakeBaseFee`)

* TestStateProcessorErrors: Use base fee of three

This was the case before enabling espresso and setting the default faked
base fee to zero.

* odr_test: use MockSysContractCallCtx

* Use NewTransaction for eth compatible tx

The goal of this commit is to reduce the diff to the original code base.
To achieve this, the naming of the new transaction functions is changed:

* Create an Ethereum compatible transaction:
  before: NewTransactionEthCompatible
  after: NewTransaction (like in geth)
* Create an transaction with Celo fields:
  before: NewTransaction
  after: NewCeloTransaction

Analogously, `NewContractCreation` defaults to Ethereum compatible transactions.

When going through the code base to fix the compilation errors introduced by these name changes, I made the decision by the following criteria:
* If Celo specific tx fields are used, use NewCeloTransaction
* Otherwise, try to use NewTransaction to minimize the diff compared to geth
* If this leads to a failing test, use NewCeloTransaction instead. This might not be the optimal choice in every case, but since this commit is already quite large, this helps to keep the changes very simple.

* Add GFork to chainConfig deepCopy

---------

Co-authored-by: Gaston Ponti <ponti@clabs.co>
Co-authored-by: Eela Nagaraj <eela@clabs.co>