Hex encode difficulty value in JSON-RPC
We add the value to a map, so our usual struct encoding steps do not apply and we have to explicitly ensure that it is hex encoded (as should be the case for all [quantities returned in JSON RPC](https://ethereum.org/en/developers/docs/apis/json-rpc/#quantities-encoding)).

We missed this because the RPC is only e2e tested using Ethers and that
consumed the plain number without complaint.