Eth header hash compatibility (#2130)
* Reorder header fields for Eth compatibility

* Use custom RLP Header serialization

To maintain compatibility with blocks from before Gingerbread, we need
to use the same serialization as back then (when the fields were not
present). This also ensures we keep the same header hashes, as those are
based on the RLP representation.

* Allow missing fields in JSON Header

These are not present in older blocks and we want to be able to decode
both the old and new blocks.

* Minor tweaks for e2e test

* Rename test to TestEthersJSCompatibilityDisableAfterGingerbread, so
  that `go test -run` can match that name without matching others
* Check for sha3Uncles value

* Add Ethereum hash compatibility test

* Update monorepo commit

Required change:
* Update header hashing for eth compat

* Use Header.SanityCheck from geth

Not that we have reintroduces the removed header fields, we can check
them in the SanityCheck, again.

* Set new header fields in genesis, too

The genesis handling can be brought close to geth than this, but this is
already a step into the right direction.

* Add case to test Celo header hash compat

with blocks before Gingerbread.