Revert "fix: go fmt"
This reverts commit 3047321ef6812e2251c2b0c5098f2b21ec70a69e.

The format change is nothing we want to include at the moment, because it increases the diff to the upstream geth version and will cause additional merge conflicts when updating to more recent upstream versions. The code is already formatted with gofmt, but with version 1.18 specifically to avoid this problem. You probably used a more recent gofmt version.