fix: udp already listening issue (#2147)
Main change:
* fix: listen udp already in use

Undone changes:
* fix: go fmt
* Revert "fix: go fmt"

The format change is nothing we want to include at the moment, because it increases the diff to the upstream geth version and will cause additional merge conflicts when updating to more recent upstream versions. The code is already formatted with gofmt, but with version 1.18 specifically to avoid this problem. You probably used a more recent gofmt version.

---------

Co-authored-by: Karl Bartel <karl.bartel@clabs.co>