Handle failing fee conversion (#2159)
* Define ToCeloFn type

The declaration is repeated multiple times and I am about to change it,
so it is nice to pull it into a typedef.

* Return errors from toCeloFn

https://github.com/celo-org/celo-blockchain/issues/2134 leads us to
believe that there are failures happening in production. This change
* Handles error cases by skipping processing of respective txs instead
  of segfaulting
* Logs tx information in these cases to better understand why this is
  happening

Skipping transactions where conversion is not possible is analogous to
handling other cases of bad transactions. But since the linked issue
only happens for some nodes, I am not sure if this is the right thing to
do or if we should intentionally panic in that case to avoid harder to
debug consensus failures due to having different nodes process a
different set of transactions.