Allow setting gingerbread in genesis block
Originally it was not possible to set the baseFeeOpCodeActivationBlock
to zero in the GasPriceMinimum contract but the contract was later
updated to allow this.

This means we can set the gingerbread block to be 0 and consequently
remove the check we had in our test infrastructure that prevented
setting the gingerbread block to zero.