Fix genesis block gingerbread fields (#2245)
This PR fixes the way the the gasLimit and gasPriceMinimum are retrieved from state such that for all non genesis blocks the value returned is the value present in the state of the parent block, and for the genesis block the value returned is the value present in the state of the genesis block (because the genesis has no parent). Prior to this fix requesting the gasLimit or gasPrice minimum for the genesis block would result in an error because we would try to get the state for the genesis block's parent which of course did not exist.

It also removes a restriction in the mycelo genesis package (used by e2e tests) that prevented us from setting the gingerbread activation block to zero. This restriction was in place because originally it was not possible to set the baseFeeOpCodeActivationBlock (AKA the gingerbread block) to zero in the GasPriceMinimum contract but the contract was later updated to allow this.

It also updates all e2e tests that were setting the gingerbread block to 1 to now use 0.

These fixes fix a number of flakily failing eth compatibility and gingerbread tests, that would fail if run on the genesis block.

 TestSendCelo
 TestTraceSendCeloViaGoldToken
 TestCallTraceTransactionNativeTransfer
 TestPrestateTransactionNativeTransfer
 TestEpochBlockMarshaling
 TestStartStopValidators
 TestBlockTracingConcurrentMapAccess
 TestBlockTracingSequentialAccess
 TestRPCDynamicTxGasPriceWithBigFeeCap
 TestRPCDynamicTxGasPriceWithState
 TestRPCDynamicTxGasPriceWithoutState
 TestEthersJSCompatibility
 TestEthersJSCompatibilityDisableAfterGingerbread
 TestTransferCELO
 TestTransferERC20