Hardcode pre gingerbread block gas limits for existing chains #2 (#2229)
Ensures that for mainnet, alfajores and baklava, nodes will return the correct block gas limit up to the gingerbread fork even if they are not archive nodes when running in eth compatibility mode (I.E. --disablerpcethcompatibility is not set). More detail can be found here: https://github.com/celo-org/celo-blockchain/issues/2214

Also includes a small re-factor of GetBlockByNumber and GetBlockByHash to reduce the depth of nesting.

This change is not backwards compatible since it is changing what is returned by the rpc for historical blocks.