Allow use of finalized tag in rpc requests
This PR allows the use of the "finalized" tag to retrieve blocks or
headers via the RPC. Using the "finalized" tag is equivalent to using
the "latest" tag because all blocks are final in celo.

This is to supprt integration with chainlink since their system depends
on retrieving finalized blocks.